Hide launcher after the plugin is closed
