+++ plugiamo/src/special/assessment/cart/index.js
@@ -1,10 +1,10 @@
+import AppBase from 'app/base'
 import data from 'special/assessment/data'
 import getFrekklsConfig from 'frekkls-config'
 import Launcher from 'app/launcher'
 import mixpanel from 'ext/mixpanel'
 import withChatActions from 'ext/recompose/with-chat-actions'
 import withHotkeys, { escapeKey } from 'ext/recompose/with-hotkeys'
-import { AppBase } from 'app'
 import { branch, compose, lifecycle, renderNothing, withHandlers, withProps, withState } from 'recompose'
 import { fetchProducts, recommendedProducts } from 'special/assessment/utils'
 import { getScrollbarWidth, isSmall } from 'utils'
