+++ plugiamo/src/special/spotahome/index.js
@@ -1,8 +1,8 @@
+import AppBase from 'app/base'
 import Base from './base'
 import Launcher from 'app/launcher'
 import mixpanel from 'ext/mixpanel'
 import withHotkeys, { escapeKey } from 'ext/recompose/with-hotkeys'
-import { AppBase } from 'app'
 import { branch, compose, lifecycle, renderNothing, withHandlers, withProps, withState } from 'recompose'
 import { gql } from 'ext/recompose/graphql'
 import { h } from 'preact'
